test combining elemental user objects and aux kernel loops:
This is a failed experiment.  The loops cannot be combined due to
"aggregate" dependencies between them.  For example,
ElementIntegralUserObject accumulates state from looping over the entire
mesh.  Auxiliary kernels can then depend on this accumulated state.
Other user objects can depend on auxiliary kernels in a similar way.
This prevents the loops from being combined into a per-element stack
because the accumulated state for one (e.g. user objects) must be
determined across the entire mesh before starting the next - i.e.
computation of the two different types of objects cannot be interleaved.

This commit fails 16 of the moose tests - which isn't bad.  But the ones
that do are broken in the fundamental way described in the paragraph
above that just *can't* be fixed without falling back to separate loops
over the mesh.  And BISON only fails 21 tests.  So, moving forward, this
kind of thing is feasible, but it needs to be a separate feature/API
that users can opt-in to using designing their applications from the
start to have deliberate (or no) mesh-aggregate dependencies between
computed values.