relax check on checkpoint file presence
This is necessary because checkpoint files will be in different
locations and named differently depending on mesh splitting and
processor number.  The naive file check we did before won't work.  So
for now we skip the check for cpr/cpa files.

This is necessary for an upcoming libmesh PR that is necessary for the
upcoming #10623 moose PR.