Add cli flag to moose for formatting input files.
Use the "--format" flag to format a moose input file.  The file to
format is specified like normal via the "-i [input-file]" flag.  The
file is modified in-place.  This walks up the directory tree starting in
the current working directory looking for a "hit.fmt" file that
specifies the hit style to format in.  If none is found, a default
canonical format is used.

Also add a "--warn-format" flag to print a warning if an input file does
not comply with the canonical HIT syntax style.