add relative path node for merge operation
The merge operation needs to be more flexible and work correctly for
cases where the src and dst nodes are not root nodes in their respective
trees.  This requires the ability to generate a relative path between a
node higher in the tree and one lower (currently no need to be able to
do multi-directional traversals - i.e. generate relative paths involving
'../' stuff). We add the Node::pathfrom function to accomplish this.
The merge operation then can use it to generate paths between it's walk
operation's start node and the current node being walked over.