Fix bug with multiple material objects providing same property.
When:

* multiple material objects provide the same property via block restriction
* one or more of the materials is a non-AD object
* one or more of the materials is an AD object
* we are not operating on the first non-linear iteration
* one or more objects retrieve that material property as an AD property.

there was a "race" of sorts to which object's declareProperty call ran
first.  If the AD object's declare ran first, then _use_ad was marked as
true in that object.  When the property was computed in the part of the
domain where the non-AD object operated, it was stored in the _value
parameter of a MooseADWrapper object. Then when another object that used
the property as an AD property (i.e. via getADMaterialProperty), this
would trigger a call to the MooseADWrapper dn() function that normally
had an if condition that would fire true to cause the wrapper's
_dual_number member to be synchronized to its _val member that had just
been calculated by the declarer.  But in this case, because of the
_use_ad race, this if condition was marked false, and the _dual_number
was not synchronized. This caused the property user to see an old
garbage value for the material property rather then the one that had
just been computed by the non-AD material object.